+++ src/net/sourceforge/kolmafia/request/AdventureRequest.java
@@ -1219,7 +1219,7 @@ public class AdventureRequest extends GenericRequest {
     return false;
   }
 
-  public static int getAdventuresUsed(final String urlString) {
+  public static long getAdventuresUsed(final String urlString) {
     KoLAdventure adventure = AdventureDatabase.getAdventureByURL(urlString);
     return adventure == null ? 0 : adventure.getRequest().getAdventuresUsed();
   }
